MUSIC DATA ANALYSIS
This project is on any song applications which we have seen in the market like gaana.com, wink, apple music and so on. We would fetch the data from this application to perform analysis like to know how many songs are heard by user and liked by the user, how many songs were completely listened by the user and how many were terminated in between and so on.